StartEntities() method ran before entities exist. so i made an event that fires on entity creation. made timer service able to have per entity timers. now i have health regen

+++ Main/Shared/Blocks/HealthBlock.cs
@@ -9,5 +9,5 @@ public class HealthBlock : Block
     [Replicated(ReplicationMode.Reliable)] public int Health = 100;
     [Replicated(ReplicationMode.Reliable)] public int MaxHealth = 100;
     [Replicated(ReplicationMode.Reliable)] public int HealthRegen = 1;
-    public double RegenRate = 0.5;
+    public float RegenRate = 0.5f;
 }
